Add UI for filtering Case Contact Report [#894]
Adds support for filter Case Contact Reports by 7 fields

* Assigned to supervisors (choose from multi-select listbox)
* Volunteers a.k.a created_by (choose from multi-select listbox)
* Contact type (choose from multi-select listbox)
* contact_type_group (choose from multi-select listbox)
* Wants driving reimbursement (two checkboxes: yes / no)
* Contact made (two checkboxes: yes / no)
* Transition aged youth (two checkboxes: yes/no)

While this commit follows the issue, we observed that filtering by case
contact booleans with a yes / no set of checkboxes creates a UX problem.

When a user select Yes, those case contacts are included.
When a user selects No, those case contacts are included.

When a user selects both Yes and No, the behavior is not defined.
When a user selects neither, noth yes and no values are selected.

An alternative, and perhaps better UX alternative is to have 3 radio
buttons for each boolean, that include yes, no, and both witha default
to both.

This seems more likely to be the filtering behavior being requested
without the confusion of what will happen if I were not to check each or
all at once.  With the radio buttons, it is clear a single action can
occur. With check boxes, there are 4 actions, for only 3 outcomes
defined.

Co-authored-by: Polly <pollysileo@gmail.com>
Co-authored-by: Jim Nanney <jnanney@gmail.com>